Fix: Move from checker to replacer
